Create start.bat
If you've built a .jar artifact to the default location in your out folder, running this in IntelliJ or running it with command prompt **from file explorer**, this will open that .jar.

Opening this with "Open In > Associated App" will open command prompt but will not run the command from the correct directory.