Method Creations, Cleanup, Verbosity
## Coordinate.java
Coordinate now has an overloaded `isEmpty()` method.
- With no parameters, it only checks for entities and items.
- With an array of Tiles, it will check to see if a given tile matches any of them in the array.

## Manager.java
Fixed a bug where denying the custom map meant a default map never activated.

## Map.java
Player is now notified when moving to an unmovable coordinate when Verbose logging is false.

If moving throws an index out of bounds exception, the player is told that it is out of bounds regardless of verbosity.

## Tile.java
Now has an equals() and hashCode() method

## CustomMap.java
Some cleanup and testing

## Debug.java
Cleanup

## Verbose.java
Only shows an error if verbosity is true.