Position Storage and Coordinate Changes
THIS IS NOT FINISHED.

In summary, `GameObject`s now store their position, and collections of `GameObject`s **are now dependent on the map, NOT the coordinate.**

This frees up a ton of memory as coordinates still "stored" a lot of info in those collections, even the ones that were empty.

In addition, symbols have changed to Strings instead of characters and I plan to change how the map prints and updates entirely.

Once functional parity with main is reached--and assuming I like it better this way--I will merge this branch back to main. Until then, wish me luck!