Inventory File Fix
The inventory file now works in Command Prompt. Before, the FileWriter would error due to not finding the file. By removing directories from the path, it now works.

This does mean that the inventory file is now in the project root folder. Since what the inventory file contains is redundant for development purposes [and instead Manager.displayInventory() should be referenced], inventory.txt is gitignored.