About the last 5 commits....
The migration to Gradle went....rockier than expected. The change was jarring enough -- and I am inexperienced enough -- that I reverted every change back to and including the migration.

The name ConsoleDungeon is staying, however.

Originally, my src folder somehow became a package, but now that that is resolved after remigrating, it should be smooth from here.